chore: clear some unneeded code may add later when needed
